pyproject.toml
[project]
name = "r2x"
version = "2.0.0"
authors = [
    { name = "pesap", email = "pesap@users.noreply.github.com" },
    { name = "kodiobika", email = "kodiobika@users.noreply.github.com" },
    { name = "Vivienne Liu", email = "vivienne.liu@nrel.gov"},
    { name = "Micah Webb", email = "micahpw@user.noreply.github.com"},
    { name = "Kinshuk Panda", email = "7632244+k1nshuk@users.noreply.github.com"},
    { name = "mcllerena", email = "mcllerena@users.noreply.github.com"},
    { name = "jarry7", email = "27745389+jarry7@users.noreply.github.com"},
]
maintainers = [
    { name = "pesap", email = "pesap@users.noreply.github.com" },
    { name = "Kinshuk Panda", email = "7632244+k1nshuk@users.noreply.github.com"},
    { name = "mcllerena", email = "mcllerena@users.noreply.github.com"},
    { name = "jarry7", email = "27745389+jarry7@users.noreply.github.com"},
    { name = "kodiobika", email = "kodiobika@users.noreply.github.com" },
]
description = "Translation Plugins for mainstream models."
readme = "README.md"
license = {file = "LICENSE.txt"}
requires-python = ">=3.11"
keywords = [
    "power-systems",
    "energy-modeling",
    "model-translation",
    "reeds",
    "plexos",
    "switch",
    "sienna",
    "plugin-architecture",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: File Formats :: JSON",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "r2x-sienna",
    "r2x-plexos",
    "r2x-reeds",
    "r2x-core",
    "pyyaml",
    "loguru",
]

[project.urls]
Homepage = "https://github.com/NREL/r2x"
Documentation = "https://nrel.github.io/r2x/"
Repository = "https://github.com/NREL/r2x"
Issues = "https://github.com/NREL/r2x/issues"
Changelog = "https://nrel.github.io/r2x/CHANGELOG.html"

[project.scripts]
plexos_to_sienna = "r2x_core.cli:run_modifier"
validate_plexos = "r2x_core.cli:run_modifier"

[dependency-groups]
dev = [
    "h5py-stubs>=0.1.2",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-coverage>=0.0",
    "ruff>=0.13.2",
    "ty>=0.0.1a23",
]
docs = [
    "autodoc-pydantic",
    "docstr-coverage>=2.3.2",
    "furo>=2025.9.25",
    "ghp-import",
    "myst-parser",
    "sphinx>=8.0.0",
    "sphinx-copybutton",
    "sphinx-reports",
    "sphinx-tabs",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[tool.mypy]
strict = true
exclude = ["tests/"]
plugins = [
  "pydantic.mypy"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
    "integration: integration tests requiring real data (deselect with '-m \"not integration\"')",
]

[tool.ruff]
line-length = 110
extend-exclude = [
    ".venv",
    "venv",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "D",      # pyddoc
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "PIE",    # flake8-pie
    "PERF",   # perflint (performance)
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B027",   # Empty method in abstract class without @abstractmethod (intentional for hooks)
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
# Ignore `D` rules everywhere except for the `src/` directory.
"!src/**.py" = ["D", "S101"]

[tool.ruff.lint.isort]
known-first-party = ["r2x_core"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.coverage.run]
omit = [
    "tests/*",
    ]

[tool.setuptools]
packages = ["r2x"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
r2x = [
    "translations/plexos_to_sienna/config/*.json",
]

[tool.uv.sources]
r2x-plexos = { git = "git+ssh://git@github.com/NREL/r2x-plexos", rev = "main" }
r2x-sienna = { git = "git+ssh://git@github.nrel.gov/PCM/r2x-sienna", rev = "main" }
r2x-core = { path = "../../r2x-core/dev", editable = true }
